requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and async
httpx==0.27.0
aiohttp==3.9.1

# API server
fastapi==0.115.0
uvicorn[standard]==0.30.0

# HTML parsing
beautifulsoup4==4.12.2
lxml>=5.0.0

# Tech stack detection
python-Wappalyzer==0.3.1

# Container orchestration
podman==4.9.0
docker==7.0.0

# Database
supabase==2.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# OpenAI
openai==1.6.1

# Vulnerability databases
nvdlib==0.7.4
requests==2.31.0

# Security and cryptography
pyjwt==2.8.0
cryptography==41.0.7

# Linting
ruff==0.8.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Reporting
jinja2==3.1.2
markdown==3.5.1

# Utilities
python-dateutil==2.8.2
